Team(CF-ProbSet)(Kumail).cpp

+++ Team(CF-ProbSet)(Kumail).cpp
@@ -4,11 +4,11 @@ using namespace std;
 
 int main()
 {
-    int p;
-    int choice[3];
-    int solP = 0;
+    short p;
+    short choice[3];
+    short solP = 0;
     cin>>p;
-    for(int i = 0; i < p; i++)
+    for(short i = 0; i < p; i++)
     {
         cin>>choice[1]>>choice[2]>>choice[3];
         if(choice[1]+choice[2]+choice[3] > 1)
